requirements.txt
piexif==1.1.3
Pillow>=10.0.0
pywin32>=300
ttkbootstrap>=1.10.0
pyinstaller>=6.0.0
